+++ src/symmetrykit/skparser.cpp
@@ -51,7 +51,7 @@ std::vector<std::tuple<double3, std::size_t, double>> SKParser::firstTestFrame()
       for (const std::shared_ptr<SKAsymmetricAtom>& atom : structure->atoms)
       {
         std::tuple<double3, std::size_t, double> atomTuple = std::make_tuple<double3, std::size_t, double>(
-            atom->position(), static_cast<std::size_t>(atom->elementIdentifier()), 1.0);
+            atom->position(), atom->elementIdentifier(), 1.0);
         atoms.push_back(atomTuple);
       }
     }
